==> src/5-data-product-description.ipynb <==
# Describe data products
1. Synthetic population attributes
2. Activity plans
3. Agents' moving trajectories
4. Road network

%load_ext autoreload
%autoreload 2
%cd D:\sysmo-data-pt
%matplotlib inline

from tqdm.notebook import tqdm
import pandas as pd
import pickle
import os
os.environ['USE_PYGEOS'] = '0'
import geopandas as gpd

## 1. Synthetic agents

df_pop = pd.read_parquet('dbs/product/syn_pop_all.parquet')
print(df_pop.head())

df_pop.groupby('feasibility').size()

## 2. Activity plans

df_plans = pd.read_parquet('dbs/product/plans_0.parquet')
print(df_plans.head())

## 3. Agents' moving trajectories

df_traj = pd.read_parquet('dbs/product/events_0.parquet')
print(df_traj.head())

## 4. Road network 
With geometry corresponding to 'link' column in the trajectories.

gdf_network = gpd.read_file('dbs/product/network.shp')
print(gdf_network.head())
